Visual Effects
Updated to unity 2020. I thought it was updated when I started the
project, but guess not.
Added URP to get shader graph

Added some shaders to adjust the saturation, value, and alpha of ores.

The layers move when you change layers. The near one moves closer to the
camera as the player moves away from the camera. Gives a bit more space
between the layers.

Camera follows the player around. The camera gets closer to the player the
deeper they are. If the camera movement is too slow in the z direction, it
feels a bit sickening with the different rates things are moving.

Shader ramblings
The alpha changes when the camera gets nearer. It's not really perfect.
The camera moves at a different rate than the layers of rock do.
Camera moves from -3.7 to 1.3, Rock moves from 0 to 4. But then the rock
layers move a bit more based on what layer the camera is on.
All of this compounded with the camera moving when the player does his
mining 'animation'. The 'animation' is really where the trouble is I
think. Everything else is working, but very finnicky. Change 1 thing and I
have to change 4 other things in different areas.